Remove idea about porting TagLibSharp.
The idea is OK, and I have managed to get it to compile for .NET Core &
Standard. However, it would require a lot of work to modernise the
code-base. This is something I am not particularly interested in doing
so re-using another library (or even just something like ffprobe with
formatted output) is likely to be more useful to me.